+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * **Breaking change**: Change the `lines` properties of type `List<String>` to
   `List<Line>`.
 * Add a new syntax `HtmlBlockSyntax` to parse HTML blocks.
+* Add an `enableTagfilter` option to `HtmlRenderer` to eanble GFM `tagfilter`
+  extension.
 * Add a new syntax `DecodeHtmlSyntax` to decode HTML entity and numeric
   character references.
 * Add a new syntax `SoftLineBreakSyntax` to remove the single space before the